refactor(i18n): 번역 콘텐츠 type 상수화
