+++ src/app/[lang]/(editor)/experience/page.tsx
@@ -1,5 +1,5 @@
 import ExperienceFormView from '@/features/Experience/ExperienceFormView';
-import { type LangParams } from '@/i18n/config';
+import { type LangParams } from '@/i18n/types';
 import { getDictionary } from '@/i18n/utils';
 
 const Page = async ({ params: { lang } }: LangParams) => {
