+++ src/app/[lang]/page.tsx
@@ -1,7 +1,7 @@
 import Footer from '@/features/Home/Footer';
 import Header from '@/features/Home/Header';
 import Main from '@/features/Home/Main';
-import { type LangParams } from '@/i18n/config';
+import { type LangParams } from '@/i18n/types';
 import { getDictionary } from '@/i18n/utils';
 
 export default async function Home({ params: { lang } }: LangParams) {
